Move access control to /auth
