+++ apps/web/pages/space/[id]/[entityId].tsx
@@ -3,15 +3,15 @@ import { Params } from '~/modules/params';
 import { Network } from '~/modules/services/network';
 import { StorageClient } from '~/modules/services/storage';
 import { Triple } from '~/modules/types';
-import { getEntityName } from '~/modules/utils';
 import { LinkedEntityGroup } from '~/modules/components/entity/types';
 import { ReadableEntityPage } from '~/modules/components/entity/readable-entity-page';
-import { useAccessControl } from '~/modules/state/use-access-control';
+import { useAccessControl } from '~/modules/auth/use-access-control';
 import { useEditable } from '~/modules/state/use-editable';
 import { EditableEntityPage } from '~/modules/components/entity/editable-entity-page';
 import { EntityStoreProvider } from '~/modules/state/entity-store-provider';
 import { useEffect } from 'react';
 import { usePageName } from '~/modules/state/use-page-name';
+import { Entity } from '~/modules/entity';
 
 interface Props {
   triples: Triple[];
@@ -94,7 +94,7 @@ export const getServerSideProps: GetServerSideProps<Props> = async context => {
     props: {
       triples: entity.triples,
       id: entityId,
-      name: getEntityName(entity.triples) ?? entityId,
+      name: Entity.name(entity.triples) ?? entityId,
       space,
       linkedEntities,
       key: entityId,
